move PyQt interface to separate module
